+++ riscv-macros/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- New `entry` macro for the Rust entry point (for `riscv-rt`).
+- New `rvrt-u-boot` feature to adapt `entry` macro for U-Boot.
 - New `post_init` macro for Rust routines that must be executed before main.
 - New `rvrt_llvm_arch_patch` and `rvrt_default_start_trap` for generating assembly code
   required by the `riscv-rt` crate.